Try/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
groq
requests
pillow
python-multipart